+++ src/backend/InvenTree/InvenTree/api_version.py
@@ -1,10 +1,13 @@
 """InvenTree API version information."""
 
 # InvenTree API version
-INVENTREE_API_VERSION = 535
+INVENTREE_API_VERSION = 536
 """Increment this API version number whenever there is a significant change to the API that any clients need to know about."""
 
 INVENTREE_API_TEXT = """
+
+v536 -> 2026-08-30 : https://github.com/inventree/InvenTree/pull/xxxx
+    - Adds SCIM 2 provisioning support
 
 v535 -> 2026-08-29 : https://github.com/inventree/InvenTree/pull/12739
     - Adds SSO registration API endpoint for creating new users via SSO